Delete features moved to the SonarQube extension
