+++ Tasks/Common/codeanalysis-common/gradlesonar.ts
@@ -5,55 +5,34 @@ import fs = require('fs');
 import tl = require('vsts-task-lib/task');
 import trm = require('vsts-task-lib/toolrunner');
 
-// Lowercased file names are to lessen the likelihood of xplat issues
-import sqCommon = require('./SonarQube/common');
-
 // Apply arguments to enable SonarQube analysis.
 // Returns the changed toolRunner. Has no effect if SonarQube is not enabled.
 export function applyEnabledSonarQubeArguments(gradleRun: trm.ToolRunner | any): trm.ToolRunner | any {
-    if (!sqCommon.isSonarQubeAnalysisEnabled()) {
-        return gradleRun;
-    }
 
     // #1: Inject custom script to the Gradle build, triggering a SonarQube run
     // Add a custom initialisation script to the Gradle run that will apply the SonarQube plugin and task
-    // Set the SonarQube Gralde plugin version in the script
-    let initScriptPath: string = path.join(__dirname, 'sonar.gradle');
-    let pluginVersion: string = sqCommon.getSonarQubeGradlePluginVersion();
-    let scriptContents: string= fs.readFileSync(initScriptPath, 'utf8');
-    scriptContents = scriptContents.replace('SONARQUBE_GRADLE_PLUGIN_VERSION', pluginVersion);
-    tl.writeFile(initScriptPath, scriptContents);
+    // Set the SonarQube Gradle plugin version in the script
+    let pluginVersion: string = getSonarQubeGradlePluginVersion();
+    if (pluginVersion !== 'applied') {
+        let initScriptPath: string = path.join(__dirname, 'sonar.gradle');
+        let scriptContents: string= fs.readFileSync(initScriptPath, 'utf8');
+        scriptContents = scriptContents.replace('SONARQUBE_GRADLE_PLUGIN_VERSION', pluginVersion);
+        tl.writeFile(initScriptPath, scriptContents);
+        // Specify that the build should run the init script
+        gradleRun.arg(['-I', initScriptPath]);
+    }
 
-    // Specify that the build should run the init script
     gradleRun.arg(['sonarqube']);
-    gradleRun.arg(['-I', initScriptPath]);
-
-    // #2: Configure additional command-line parameters
-    // Add parameters to connect to the SonarQube server for reporting
-    gradleRun = sqCommon.applySonarQubeParameters(gradleRun);
 
     return gradleRun;
 }
 
-// Points SonarQube to the CC file as it is in a non-standard location. Not required for Jacoco.
-// Returns the changed toolRunner. Has no effect if SonarQube or code coverage is not enabled.
-export function applySonarQubeCodeCoverageArguments(gradleRun: trm.ToolRunner | any, isCodeCoverageEnabled: boolean, ccTool: string, reportPath: string): trm.ToolRunner | any{
-    // Apply relevant arguments if both SonarQube and code coverage features are enabled
-    if (sqCommon.isSonarQubeAnalysisEnabled() && isCodeCoverageEnabled && ccTool.toLowerCase() === 'cobertura' && reportPath ) {
-        gradleRun.arg('-Dsonar.cobertura.reportPath=' + reportPath);
+function getSonarQubeGradlePluginVersion(): string {
+    let pluginVersion = '2.6.1';
+    let userSpecifiedVersion = tl.getInput('sqGradlePluginVersion');
+    if (userSpecifiedVersion) {
+        pluginVersion = userSpecifiedVersion.trim();
     }
-
-    return gradleRun;
+    return pluginVersion;
 }
 
-// Upload a build summary with links to available SonarQube dashboards for further analysis details.
-export function processSonarQubeIntegration(): Q.Promise<void> {
-    if (!sqCommon.isSonarQubeAnalysisEnabled()) {
-        return Q.when();
-    }
-
-      // the output folder may not be directly in the build root, for example if the entire project is in a top-level dir
-    let reportTaskGlob: string = path.join(tl.getVariable('System.DefaultWorkingDirectory'), '**', 'build', 'sonar', 'report-task.txt');
-
-    return sqCommon.processSonarQubeIntegration(reportTaskGlob);
-}
